requirements.txt
openai==2.30.0
scikit-learn==1.8.0
python-dotenv==1.2.2
notebook==7.5.5
ipykernel==7.2.0
matplotlib==3.10.8
